+++ common/src/main/java/net/creeperhost/resourcefulcreepers/BuildInfo.java
@@ -1,8 +1,8 @@
 package net.creeperhost.resourcefulcreepers;
 
 public class BuildInfo {
-    public static final String version = "3b114170d7";
-    public static final String buildTimestamp = "Thu Nov 24 13:34:59 GMT 2022";
+    public static final String version = "ecb13000ac";
+    public static final String buildTimestamp = "Thu Nov 24 14:06:48 GMT 2022";
 
     public String toString() {
         return "version         : " + version + "\n" +